[JENKINS-56525] Add filters for roles and users
With the new filters managing of roles and user will become easier.
On the manage roles pages when adding more than 10 roles, the filter is
shown and when getting below 10 it is hidden.
On the assign roles page the input to filter roles will be enabled when
more than 20 roles are configured.
The input to filter users/groups will be enabled when more than 10 users
are added. The input is dynamic and will automatically appear when
adding the 11th user/group.
The filters are case insensitive