speed up user checking on assign roles
The user checking is done sequentially by the FormChecker from core.
With many users this checking can take long time (reports are over 10
min for 300 users). This also depends on the securityRealm performance
and how many of these users really exist.
Confirmed with an AD and 300 users to take 10 to 12 min, with about 60%
of users not existing anymore.

To get a recognizable performance gain http2 needs to be enabled. When
this is done and running with 50 max parallel, the above scenario took
only around 1 min to load. Without http2, with 10 parallel connections
loading time went down to 5 min, with no further improvement with more
connections. It can even have bad impact in chrome, other browser
windows accesing the root of Jenkins was hanging with "waiting for
socket" in chrome.