[JENKINS-19934] fix project naming strategy (#179)
* [JENKINS-19934] WIP fix project naming strategy

The project naming strategy is currently not considering if the user
has the permission to create a job with a given name at all. I just
checks if the passed name matches any of the patterns.
Also it doesn't consider the full name, e.g. when a project is created
inside a folder.
This can lead to the situtation that a user creates a job but is then
unable to configure the job.

This change will do the following:
Prerequisites:
 - Rolebased project naming strategy is enabled.
 - for each role when create is enabled also configure and read should
   be enabled

If a user has either globally or in any item role the create permission
then he will see the "New item" link in the side panel.
A user that has global create permissions can create jobs with any name.
For a user that has create permission on one or more item roles the
entered name is matched against the role and the users permissions.

Current limitation is that it only works reliably via the UI. When
trying to create a job via the CLI, there is no staplerrequest that can
be used to find the parent item. So job creation via cli will fall back
to the old behaviour and just check the item name itself.
In case JENKINS-68602 gets resolved requests coming in via the CLI would
also be properly checked.

This also adds an admin monitor that warns when the role based project
naming strategy is not enabled.

* fix javadoc ref

* add more tests

* fix group resolution

can't use the code path from authorities as roles that is wrong.
Add tests that verify group permissions work properly

* fix javadoc

* code cleanup

* optimize checking create permission

* update documentation

* remove unused method

* fix formatting

* exclude macro roles from naming check

* add tests

* adjust readme