Role Generator
The Role Generator allows to generate roles that are a combination of
one or more permission templates and one or more role templates.
A permission template defines a name with a set of permissions.
A Role template defines a name with a pattern.
The generator now applies each permission template to each role template
creating a new Role with name "#<permissiontemplate name>-<role template
name>
E.g. "#admin-team1"