delete performance folder
the performance folder was a first try to demonstrate the performance
issues of the plugin.
The config that is maintained there (plugins.txt, poms, ..) would need
constant maintenance and frequently subject to changes of dependabot.
Since then jmh tests have been introduced that allow performance
measurements via maven.
So deleting this will help to reduce noise.