implicit view permissions
Do not check if the sid has anywhere view permissions. This doesn't work
when on global level a group grants access and on item level only a user
or vice versa.
When this macro is used the View.READ permission is implied. Jenkins
itself will show only those views, that contain at least one item.