remove prototype, remove inline java script
prototype.js is part of core, but this is not maintained anymore. There
are activities going on in core to remove all usages of this js library

Content-Security policy demands that there is no inline javascript in
html files. So move all code from the jelly files to dedicated js files.