[JENKINS-71138] use sidepanel for easier navigation
So far the sidepanel from the main page was used and there was an own
index page just linking the manage, assign and macros pages. This means
there was no way of directly jumping from one of the pages to the other.
One had to always go over the index page.
With an own sidepanel, we reach now directly the manage roles page from
the manage Jenkins page and can directly go the other role strategy
pages.