avoid flickering when moving in the table
the hover effect was only on the input. That lead to the highlight be
turned leaving the input end turn it on again in the next cell of the
table. The mouseover is now on the td which avoids the flickering.

move the templates to a template element out of the table.