[JENKINS-71772] Remove runtime dependency on authorize-project plugin
https://issues.jenkins.io/browse/JENKINS-71772 notes that
https://github.com/jenkinsci/role-strategy-plugin/commit/a5f27678f6d61a330597b9b157ca7468158835f5
inadvertently changed the authorize-project plugin from a test dependency
to a runtime dependency.  The code compiles correctly without the runtime
dependency and the tests pass on my Linux computer with Java 11.

https://github.com/jenkinsci/bom/pull/2340 added authorize-project plugin
to the plugin bill of materials because I didn't detect that the new
dependency was not needed at runtime.

Thanks to James Holderness for reporting the issue.