Use plugin bom for dependency version definitions
The most recent plugin bill of materials provides the authorize-project
plugin version.  Use that version rather than managing the version
drectly in the plugin pom.

Apache commons-lang3-api is available as an API plugin that is managed by
the bom.  Test with the API plugin rather than including commons-lang3-api
in the tests.  Testing with the API plugin has the added benefit that it
is using the same version that is installed in many Jenkins controllers.