[JENKINS-75868] reduce browser memory consumption with many roles/users
The memory consumption goes up to over 7 GiB when there are 400 item
roles and a one user/group for each role. In such a scenario the UI will
need to show 160000 checkboxes.
A considerable part of the memory consumption could be attributed to the
tooltips.
With such a large table also the highlighting becomes very laggy.

To improve the situation tooltips and table highlighting will be
disabled when the number of checkboxes exceeds 40000 in the item roles
table.