[JENKINS-75695] Pagination for item role assignments and UI refresh
Working with more than 50 or 100 roles was close to impossible via the
UI. Defining e.g. 400 roles and assigning one user per role resulted in a
matrix of 160k checkboxes.
While using the filters helped to improve the situation, the sheer mass
of checkboxes also impacted form submission. By default the number of
fields in a form is limited to 10000 and the total form size is 200k.
Both values are easily reached already at around 100 roles.

This changes introduces pagination on users. The form no longer passes
all checkboxes but just a json containing the assigned roles for each
user/group. This means just a single field and as usually for each user
only a few roles are checked also a small form.

The tables have been redesigned to use the modern styling of tables in
Jenkins.