properly filter the permission groups by scope
permissions are scoped to reflect for which area they apply. The current
implementation was manually removing certain groups existing in core. So
any permissions defined in plugins were shown for items and agents even
when scoped to Jenkins only (e.g. lockable resources)
And when a group contained 2 permissions with different scope those were
always shown.
With this change the item roles are based on the ITEM_GROUP scope (which
includes ITEMS and RUNS) and agent roles are based on the COMPUTER
scope. Permissions are filtered so for a group that contains a
permission for item and computer the one for item is not shown in the
agent roles.